Pass dtype through in random_color
random_color accepts and documents a dtype argument but never forwarded
it to hsv_to_rgba, so every call returned uint8 regardless: a requested
float64 came back as uint8 0-255 instead of 0.0-1.0, uint16 was not
scaled, and an unsupported dtype did not raise. hsv_to_rgba already
implements the conversion, so forward the argument. The default
(np.uint8) is unchanged on both sides, so existing behavior is
unaffected. Extends test_random_color to cover the dtype argument.

+++ trimesh/visual/color.py
@@ -801,7 +801,7 @@ def random_color(dtype: DTypeLike = np.uint8, count: Integer | None = None) -> N
     # saturation and "value" as constant
     sv = np.ones_like(hue) * 0.99
     # convert our random hue to RGBA
-    colors = hsv_to_rgba(np.column_stack((hue, sv, sv)))
+    colors = hsv_to_rgba(np.column_stack((hue, sv, sv)), dtype=dtype)
 
     # unspecified count is a single color
     if count is None:
